refactor: extract TimerTaskHandler from foreground_timer_service
